task_app/requirements.txt
Flask
Flask-SQLAlchemy
pymysql
gunicorn
python-dotenv
jsonify
Flask-JSO